+++ src/main/java/de/kevtv/kevin/minecraft_spigot_trade_system/tradeinventory/itemstacks/MoneyPlus.java
@@ -23,7 +23,7 @@ public class MoneyPlus {
         setMoneyPlusDisplayName();
     }
 
-    public int getPlayerMoneyAmount() {
+    public double getPlayerMoneyAmount() {
         return moneyAmount.getPlayerMoneyAmount();
     }
 
@@ -35,19 +35,19 @@ public class MoneyPlus {
         return moneyAmount;
     }
 
-    public void setMoney(int money) {
+    public void setMoney(double money) {
         moneyAmount.setMoney(money);
     }
 
-    public void addMoney(int moneyToAdd) {
+    public void addMoney(double moneyToAdd) {
         moneyAmount.addMoney(moneyToAdd);
     }
 
-    public void remMoney(int moneyToRem) {
+    public void remMoney(double moneyToRem) {
         moneyAmount.remMoney(moneyToRem);
     }
 
-    public int getMoney() {
+    public double getMoney() {
         return moneyAmount.getMoney();
     }
